Fix bug when there is no rice-box
Former-commit-id: aa3efbd6ac0955c49a0187851d999b48fd2fa784 [formerly 9703b344fc39b60d3d2f637baa9fa134a4700149] [formerly 04eb98512022fac1cd346600ab9780c178c18aa6 [formerly 50bb68ddc820bfbece26178553a3a214061ab17d]]
Former-commit-id: c6c70deb184efd8b59d99357fc1be6555e76a071 [formerly dee69992cb1423d7a692d9dd4df5ece0bd0262dc]
Former-commit-id: 9b35e3926819a1c0e6e8c042d73d9b089f9d3be6

+++ assets/build/dev.js
@@ -10,8 +10,13 @@ var config = require('./config')
 var webpackConfig = require('./webpack.dev.conf')
 var fs = require('fs')
 
-fs.unlinkSync('./rice-box.go')
-fs.unlinkSync('./caddy/hugo/rice-box.go')
+if (fs.existsSync('./rice-box.go')) {
+  fs.unlinkSync('./rice-box.go')
+}
+
+if (fs.existsSync('./caddy/hugo/rice-box.go')) {
+  fs.unlinkSync('./caddy/hugo/rice-box.go')
+}
 
 rm(path.join(config.assetsRoot, config.assetsSubDirectory), err => {
   if (err) throw err
